+++ cmd/exec.go
@@ -5,7 +5,7 @@ import (
 	"os"
 
 	"github.com/ahamlinman/slackbridge/internal/childproc"
-	"github.com/ahamlinman/slackbridge/slackio"
+	"github.com/ahamlinman/slackio"
 	"github.com/spf13/cobra"
 )