+++ slackio/message.go
@@ -1,9 +1,0 @@
-package slackio
-
-// Message is the type for messages received from and sent to a single Slack
-// channel.
-type Message struct {
-	ID        int
-	ChannelID string
-	Text      string
-}